+++ parsons/ngpvan/activist_codes.py
@@ -42,10 +42,26 @@ class ActivistCodes(object):
         logger.info(f'Found activist code {activist_code_id}.')
         return r
 
-    def toggle_activist_code(self, id, activist_code_id, action, id_type='vanid',
-                             contact_type_id=None, input_type_id=None, date_canvassed=None):
+    def toggle_activist_code(self, id, activist_code_id, action, id_type='vanid'):
+        # Internal method to apply/remove activist codes. Was previously a public method,
+        # but for the sake of simplicity, breaking out into two public methods.
+
+        logger.info(f'Method deprecated. Use apply_activist_code() or remove_activist_code().')
+
+        response = {"activistCodeId": activist_code_id,
+                    "action": action_parse(action),
+                    "type": "activistCode"}
+
+        r = self.apply_response(id, response, id_type)
+
+        logger.info(f'{id_type.upper()} {id} {action.capitalize()} ' +
+                    f'activist code {activist_code_id}')
+
+        return r
+
+    def apply_activist_code(self, id, activist_code_id, id_type='vanid'):
         """
-        Apply or remove an activist code to or from a person.
+        Apply an activist code to or from a person.
 
         `Args:`
             id: str
@@ -61,13 +77,24 @@ class ActivistCodes(object):
             Status code 204
         """
 
-        response = {"activistCodeId": activist_code_id,
-                    "action": action_parse(action),
-                    "type": "activistCode"}
+        return self.toggle_activist_code(id, activist_code_id, 'Apply', id_type=id_type)
 
-        r = self.apply_response(id, response, id_type)
+    def remove_activist_code(self, id, activist_code_id, id_type='vanid'):
+        """
+        Remove an activist code to or from a person.
 
-        logger.info(f'{id_type.upper()} {id} {action.capitalize()} ' +
-                    f'activist code {activist_code_id}')
+        `Args:`
+            id: str
+                A valid person id
+            activist_code_id: int
+                A valid activist code id
+            action: str
+                Either 'apply' or 'remove'
+            id_type: str
+                A known person identifier type available on this VAN instance
+                such as ``dwid``
+        Returns:
+            Status code 204
+        """
 
-        return r
+        return self.toggle_activist_code(id, activist_code_id, 'Remove', id_type=id_type)
